Started path finding algorithm

+++ RovinePerdute/src/EdgeCalculator.java
@@ -1,3 +1,3 @@
 public interface EdgeCalculator {
-    public int calculateEdgeDistance(City first, City second);
+    public double calculateEdgeDistance(City first, City second);
 }
