+++ RovinePerdute/src/Metztli.java
@@ -1,8 +1,8 @@
 public class Metztli implements EdgeCalculator {
 
     @Override
-    public int calculateEdgeDistance(City first, City second) {
-        int fuel = second.getH() - first.getH();
+    public double calculateEdgeDistance(City first, City second) {
+        double fuel = second.getH() - first.getH();
         return Math.abs(fuel);
     }
 }
